requirements.txt
openai==1.14.3
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.1.1
langchain_core==0.1.42
faiss-cpu==1.8.0
tiktoken==0.6.0
markdown2==2.4.13
python-dotenv==1.0.1
langdetect==1.0.9
unstructured[all-docs]==0.12.6
mammoth==1.7.1
markdownify==0.12.1
langchain_text_splitters==0.0.1
doc2docx==0.2.4
deep_translator==1.11.4
ruff==0.3.4
toml==0.10.2
tenacity==8.2.3
scikit-learn==1.4.0
numpy==1.26.4
dill==0.3.8
fuzzywuzzy==0.18.0
rapidfuzz==3.6.1
python_Levenshtein==0.24.0
Flask==3.0.3
Flask-PyMongo==2.3.0
certifi==2024.2.2
google-api-python-client==2.137.0
fastembed==0.3.4
fake_useragent-1.5.1